Updates to alert web component functionality
