--- src/components/Alert-wc/presets/0-default.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import AlertWc from '../Alert-wc';
-
-export default (
-  <AlertWc
-    uxpId="alert"
-    size="md"
-    state="info"
-    title=""
-    icon=""
-    type="bubble"
-    closable={false}
-    click={() => {}}
-  />
-);
