+++ src/components/Alert/Alert.js
@@ -1,189 +1,117 @@
-/* eslint-disable jsx-a11y/anchor-is-valid */
+/* eslint-disable react/no-unused-prop-types */
 import * as PropTypes from 'prop-types';
 import * as React from 'react';
-import {
-  ALERT_CLASSES,
-  BUTTON_CLASSES,
-  CLOSE_CLASS,
-  ICON_CHEVRON_RIGHT,
-  ICON_CLASS,
-  INFO_CLASS,
-  LINK_CLASSES,
-  NO_HOVER_UNDERLINE,
-  SPINNER_CLASSES,
-  UTILITY_CLASSES,
-} from '../../constants/classes';
+import { ALERT_CLASSES, DROPDOWN_CLASSES, UTILITY_CLASSES } from '../../constants/classes';
 
 export default class Alert extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      defaultIcon: 'flag',
-      showAlert: true,
-    };
+  alertRef = React.createRef();
+
+  componentDidMount() {
+    const alertElement = this.alertRef.current;
+    if (alertElement) {
+      alertElement.addEventListener('dismissAlert', this.handleDismiss);
+    }
   }
 
-  // TODO: Change name to _setIcon()
-  _setIconAndText() {
-    let defaultText = this.props.text;
-    let defaultIconName = this.state.defaultIcon;
+  componentWillUnmount() {
+    const alertElement = this.alertRef.current;
+    if (alertElement) {
+      alertElement.removeEventListener('dismissAlert', this.handleDismiss);
+    }
+  }
 
-    switch (this.props.state) {
+  handleDismiss = () => {
+    if (this.props.onDismiss) {
+      this.props.onDismiss();
+    }
+  };
+
+  renderAlertContent(defaultIconName, color, alertSize, alertType, title, defaultText) {
+    const isClosable = this.props.closable && ['bubble', 'toast'].includes(this.props.type);
+
+    return (
+      <chi-alert
+        ref={this.alertRef}
+        color={color}
+        icon={defaultIconName}
+        size={alertSize}
+        type={alertType}
+        closable={isClosable}
+        spinner={this.props.animateSpinner}
+        title={`${title}\n`}
+        class={UTILITY_CLASSES.WIDTH[100]}>
+        <span style={{ whiteSpace: 'pre-wrap' }}>{defaultText}</span>
+        {this.props.type === 'clickable' && <chi-icon icon={DROPDOWN_CLASSES.ICON_CHEVRON_RIGHT} slot={ALERT_CLASSES.CLICKABLE_ICON}></chi-icon>}
+      </chi-alert>
+    );
+  }
+
+  render() {
+    const {
+      size,
+      state,
+      title,
+      icon,
+      type,
+      active,
+      text,
+    } = this.props;
+
+    let defaultIconName;
+    let defaultText = text;
+
+    switch (state) {
       case 'success':
-        defaultIconName = 'circle-check';
+        defaultIconName = icon || 'circle-check';
         defaultText = defaultText || 'This is a success alert';
         break;
       case 'warning':
-        defaultIconName = 'warning';
+        defaultIconName = icon || 'warning';
         defaultText = defaultText || 'This is a warning alert';
         break;
       case 'danger':
-        defaultIconName = 'circle-x';
+        defaultIconName = icon || 'circle-x';
         defaultText = defaultText || 'This is a danger alert';
         break;
       case 'info':
-        defaultIconName = 'circle-info';
+        defaultIconName = icon || 'circle-info';
         defaultText = defaultText || 'This is an info alert';
         break;
       case 'muted':
-        defaultIconName = 'flag';
+        defaultIconName = icon || 'flag';
         defaultText = defaultText || 'This is a muted alert';
         break;
       case 'base':
       default:
-        defaultIconName = 'flag';
+        defaultIconName = icon || 'flag';
         defaultText = defaultText || 'This is a base alert';
         break;
     }
 
-    return { defaultIconName, defaultText };
-  }
+    const color = state;
+    const alertSize = size !== 'md' ? size : undefined;
+    const alertType = type !== 'clickable' ? type : undefined;
 
-  _alertTitle() {
-    return (
-      <p className={`${ALERT_CLASSES.TITLE} ${UTILITY_CLASSES.TEXT.LG}`}>
-        {this.props.title}
-      </p>
-    );
-  }
+    if (!active) {
+      return null;
+    }
 
-  _closableButton() {
-    return (
-      <button
-        type="button"
-        onClick={() => {
-          this.setState({ showAlert: false });
-          if (this.props.click) {
-            this.props.click();
-          }
-        }}
-        className={`
-          ${ALERT_CLASSES.CLOSE_BUTTON} 
-          ${BUTTON_CLASSES.BUTTON} 
-          ${BUTTON_CLASSES.ICON_BUTTON} 
-          ${CLOSE_CLASS}`}
-        aria-label="Close">
-        <div className={BUTTON_CLASSES.CONTENT}>
-          <i className={`${ICON_CLASS} icon-x`}></i>
-        </div>
-      </button>
-    );
-  }
+    const alertContent = this.renderAlertContent(defaultIconName, color, alertSize, alertType, title, defaultText);
 
-  _progressIcon() {
-    return (
-      <div className={ALERT_CLASSES.ICON}>
-        <svg
-          className={`${SPINNER_CLASSES.SPINNER} ${INFO_CLASS} -sm--2`}
-          viewBox="0 0 66 66">
-          <title>Loading</title>
-          <circle
-            className="path"
-            cx="33"
-            cy="33"
-            r="30"
-            fill="none"
-            strokeWidth="6">
-          </circle>
-        </svg>
-      </div>
-    );
-  }
-
-  _alert() {
-    const state = this.props.state && this.props.state !== 'base'
-      ? `-${this.props.state}`
-      : '';
-    const type = this.props.type ? `-${this.props.type}` : '';
-    const size = this.props.size ? `-${this.props.size}` : '';
-    const { defaultIconName, defaultText } = this._setIconAndText();
-    const iconName = `icon-${this.props.icon || defaultIconName}`;
-    const iconToDisplay = this.props.inProgress ? (
-      this._progressIcon()
-    ) : (
-      <i
-        className={`${ALERT_CLASSES.ICON} ${ICON_CLASS} ${iconName}`}
-        aria-hidden="true">
-      </i>
-    );
-    const title = this.props.title ? this._alertTitle() : null;
-    const alertContent = (
-      <>
-        {iconToDisplay}
-        <div className={ALERT_CLASSES.CONTENT}>
-          {title}
-          <p
-            className={ALERT_CLASSES.TEXT}
-            style={{ whiteSpace: 'pre-line' }}>
-            {defaultText}
-          </p>
-        </div>
-      </>
-    );
-    const chevronRight = (
-      <div className={ALERT_CLASSES.CLICKABLE_ICON}>
-        <i className={`${ICON_CLASS} ${ICON_CHEVRON_RIGHT}`}></i>
-      </div>
-    );
-    const closeButton = this.props.closable ? this._closableButton() : null;
-    const rightIcon = this.props.type !== 'clickable' ? closeButton : chevronRight;
-    const alert = (
-      <div
-        className={`${ALERT_CLASSES.ALERT} ${state} ${size} ${type} ${closeButton ? `${CLOSE_CLASS}` : ''}`}
-        role="alert">
+    return type === 'clickable' ? (
+      <chi-link href="#" onClick={this.props.click} no-hover-underline class={UTILITY_CLASSES.DISPLAY.BLOCK}>
         {alertContent}
-        {rightIcon}
-      </div>
+      </chi-link>
+    ) : (
+      alertContent
     );
-    const clickableAlert = (
-      <a
-        onClick={this.props.click}
-        className={`${LINK_CLASSES.LINK} ${NO_HOVER_UNDERLINE}`}
-        aria-hidden="true">
-        {alert}
-      </a>
-    );
-
-    return this.props.type !== 'clickable' ? alert : clickableAlert;
-  }
-
-  render() {
-    this._setIconAndText();
-    return this.state.showAlert ? <div ref={this.props.uxpinRef}>{this._alert()}</div> : null;
   }
 }
 
 Alert.propTypes = {
   size: PropTypes.oneOf(['sm', 'md']),
-  state: PropTypes.oneOf([
-    'base',
-    'success',
-    'warning',
-    'danger',
-    'info',
-    'muted',
-  ]),
-  inProgress: PropTypes.bool,
+  state: PropTypes.oneOf(['base', 'success', 'warning', 'danger', 'info', 'muted']),
+  animateSpinner: PropTypes.bool,
   /**
    * A textArea controller for Text
    * @uxpinpropname text
@@ -192,13 +120,24 @@ Alert.propTypes = {
   text: PropTypes.string,
   title: PropTypes.string,
   icon: PropTypes.string,
+  active: PropTypes.bool,
   type: PropTypes.oneOf(['bubble', 'toast', 'clickable']),
   closable: PropTypes.bool,
+  /**
+   * @uxpinpropname On Click
+   */
   click: PropTypes.func,
+  /**
+   * @uxpinpropname On Dismiss
+   */
+  onDismiss: PropTypes.func,
 };
 
 Alert.defaultProps = {
   size: 'md',
+  title: '',
   state: 'info',
   type: 'bubble',
+  active: true,
+  onDismiss: null,
 };
